Added categories.variables.js for datas
